Move additional stuff from 'types' to 'hostPage'
