+++ src/js/server/toolbar/langSwitch.ts
@@ -17,7 +17,7 @@
  */
 import { Observable, of as rxOf } from 'rxjs';
 
-import { HostPageEnv, IToolbarProvider } from '../../common/types';
+import { HostPageEnv, IToolbarProvider } from '../../common/hostPage';
 import { GlobalComponents } from '../../views/global';
 import { ViewUtils } from 'kombo';
 import { init as viewInit } from '../../views/langSwitchToolbar';
