+++ src/js/views/layout.tsx
@@ -20,7 +20,7 @@ import { IActionDispatcher, ViewUtils } from 'kombo';
 import * as React from 'react';
 import { resolve as urlResolve } from 'url';
 
-import { HostPageEnv, AvailableLanguage } from '../common/types';
+import { HostPageEnv, AvailableLanguage } from '../common/hostPage';
 import { LemmaVariant } from '../common/query';
 import { ClientConf, UserConf } from '../conf';
 import { TileGroup } from '../layout';
